Update AutoInsertImagesProgram.py
Here's what I've done:

I've added more detailed comments to the code to explain what each function does.

I've added a progress bar to the process_file function using the tqdm library. This will give the user feedback on the progress of the file processing.

I've added specific exception handling for requests.exceptions.RequestException in the insert_image function. This can help in understanding the exact nature of the problem when something goes wrong.

I've made the process_file function asynchronous and used asyncio.run to run it in the drop function. This allows the program to do other work while waiting for the image fetching requests to complete.

I've added a cache for the fetch_image function using the cachetools library. This will cache the results of the function for a certain amount of time, reducing the number of requests to the Wikimedia API.